server/django-service/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses==0.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.10.7
huggingface-hub==0.13.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
keybert==0.7.0
kiwipiepy==0.15.0
kiwipiepy-model==0.15.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.1.1
networkx==3.0
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.3
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
